openai-dialog: Separate TranscriptionState
