+++ services/openai-dialog/src/client.rs
@@ -1,6 +1,5 @@
 #[cfg(feature = "prompt-delay")]
 use std::collections::VecDeque;
-use std::collections::{HashMap, HashSet};
 
 use anyhow::{Context, Result, bail};
 use base64::prelude::*;
@@ -16,6 +15,7 @@ use tracing::{debug, info, trace, warn};
 #[cfg(feature = "prompt-delay")]
 use uuid::Uuid;
 
+use crate::transcription::{TranscriptionSettings, TranscriptionState};
 use crate::{Params, ServiceInputEvent, ServiceOutputEvent};
 use context_switch_core::{
     AI_AGENT_SPEAKER, AudioFormat, AudioFrame, BillingRecord, BillingSchedule,
@@ -29,51 +29,6 @@ pub struct Client {
 
     #[cfg(feature = "prompt-delay")]
     prompt_coordinator: PromptCoordinator,
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct TranscriptionSettings {
-    pub input: bool,
-    pub output: bool,
-}
-
-#[derive(Debug, Default)]
-struct TranscriptionState {
-    input_transcription_buffers: HashMap<InputTranscriptionKey, String>,
-    output_transcription_buffers: HashMap<OutputTranscriptionKey, String>,
-    output_transcription_responses: HashSet<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-struct InputTranscriptionKey {
-    item_id: String,
-    content_index: u32,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-struct OutputTranscriptionKey {
-    item_id: String,
-    output_index: u32,
-    content_index: u32,
-}
-
-impl InputTranscriptionKey {
-    fn new(item_id: String, content_index: u32) -> Self {
-        Self {
-            item_id,
-            content_index,
-        }
-    }
-}
-
-impl OutputTranscriptionKey {
-    fn new(item_id: String, output_index: u32, content_index: u32) -> Self {
-        Self {
-            item_id,
-            output_index,
-            content_index,
-        }
-    }
 }
 
 #[cfg(feature = "prompt-delay")]
@@ -491,8 +446,9 @@ impl Client {
                 },
             ) => {
                 if transcription.input {
-                    let key = InputTranscriptionKey::new(item_id, content_index);
-                    let text = self.transcription_state.apply_input_delta(key, delta);
+                    let text = self
+                        .transcription_state
+                        .apply_input_delta(item_id, content_index, delta);
                     output.text(false, text, None, None)?;
                 }
             }
@@ -506,7 +462,8 @@ impl Client {
             ) => {
                 if transcription.input
                     && let Some(text) = self.transcription_state.complete_input_transcription(
-                        InputTranscriptionKey::new(item_id, content_index),
+                        item_id,
+                        content_index,
                         transcript,
                     )
                 {
@@ -524,10 +481,9 @@ impl Client {
                 },
             ) => {
                 if transcription.output {
-                    let key = OutputTranscriptionKey::new(item_id, output_index, content_index);
                     let text = self
                         .transcription_state
-                        .apply_output_delta(response_id, key, delta);
+                        .apply_output_delta(response_id, item_id, output_index, content_index, delta);
                     output.text(false, text, None, Some(AI_AGENT_SPEAKER.into()))?;
                 }
             }
@@ -544,7 +500,9 @@ impl Client {
                 if transcription.output
                     && let Some(text) = self.transcription_state.complete_output_transcription(
                         response_id,
-                        OutputTranscriptionKey::new(item_id, output_index, content_index),
+                        item_id,
+                        output_index,
+                        content_index,
                         transcript,
                     )
                 {
@@ -741,81 +699,6 @@ impl Client {
         }
 
         Ok(())
-    }
-}
-
-impl TranscriptionState {
-    fn apply_input_delta(&mut self, key: InputTranscriptionKey, delta: String) -> String {
-        let entry = self.input_transcription_buffers.entry(key).or_default();
-        entry.push_str(&delta);
-        entry.clone()
-    }
-
-    fn complete_input_transcription(
-        &mut self,
-        key: InputTranscriptionKey,
-        transcript: String,
-    ) -> Option<String> {
-        let text = if transcript.is_empty() {
-            self.input_transcription_buffers
-                .remove(&key)
-                .unwrap_or_default()
-        } else {
-            self.input_transcription_buffers.remove(&key);
-            transcript
-        };
-
-        (!text.is_empty()).then_some(text)
-    }
-
-    fn apply_output_delta(
-        &mut self,
-        response_id: String,
-        key: OutputTranscriptionKey,
-        delta: String,
-    ) -> String {
-        self.output_transcription_responses.insert(response_id);
-        let entry = self.output_transcription_buffers.entry(key).or_default();
-        entry.push_str(&delta);
-        entry.clone()
-    }
-
-    fn complete_output_transcription(
-        &mut self,
-        response_id: String,
-        key: OutputTranscriptionKey,
-        transcript: String,
-    ) -> Option<String> {
-        self.output_transcription_responses.insert(response_id);
-        let text = if transcript.is_empty() {
-            self.output_transcription_buffers
-                .remove(&key)
-                .unwrap_or_default()
-        } else {
-            self.output_transcription_buffers.remove(&key);
-            transcript
-        };
-
-        (!text.is_empty()).then_some(text)
-    }
-
-    fn clear_output_response_tracking(&mut self, response_id: &str) {
-        self.output_transcription_responses.remove(response_id);
-    }
-
-    fn clear_item(&mut self, item_id: &str) {
-        self.input_transcription_buffers
-            .retain(|buffer_key, _| buffer_key.item_id != item_id);
-        self.output_transcription_buffers
-            .retain(|buffer_key, _| buffer_key.item_id != item_id);
-    }
-
-    fn clear_content_index(&mut self, item_id: String, content_index: u32) {
-        let key = InputTranscriptionKey::new(item_id.clone(), content_index);
-        self.input_transcription_buffers.remove(&key);
-        self.output_transcription_buffers.retain(|buffer_key, _| {
-            buffer_key.item_id != item_id || buffer_key.content_index != content_index
-        });
     }
 }
